Add cwd to subprocess calls
We were assuming this path before, which defaulted to the path you
executed django from on the host machine. The cwd was used in commands
run on inside Docker however. On Linux dev environments, this path is
most likely missing.

By specifying paths, we will get paths that have been created inside the
container.