Require conda.file when using conda in v1
Some people are migrating from pip to conda,
but they were using a v1 config,
then when migrating they are using the conda option from v2.

We are not validating this, but using conda.file is mandatory
when using the conda option (or rtd will fail without any reason).

Close #5322
Close #5228