Create a funding file
Adding this file will create a "sponsor" button on the GitHub UI for our repository. This button will open a modal which links to our external sustainability page. The PR preview shows an example.

There are a lot of possibilities here but before we spend any real time on it I'm interested to see if we get any engagement.

See:

* https://github.blog/2019-05-23-announcing-github-sponsors-a-new-way-to-contribute-to-open-source/
* https://help.github.com/en/articles/displaying-a-sponsor-button-in-your-repository