Fix tests on master
Two PRs got merged and one modified the behaviour that the
other one was depending on.

This was giving an error in the lock creation
because the base dir wasn't created yet.

The rest of tests something I forgot to change project_pk -> version_pk
(I did a grep, didn't find more)