Use different setting for footer api url
Currently we use 'api_host' in our js to call the footer api.
I'm making a separate setting just for the footer api.
This is to allow us to proxy the footer api from the same domain that
is serving the docs (this is useful for .com).

We can't proxy the whole api from the serve app because there are some
places where we call `reverse`, and on the serve app we don't have
access to all urls (like project dashboard).

We need to publish the sphinx-extension with this change first.