Use proper domain for AZURE_MEDIA_STORAGE_HOSTNAME
We can't access docs.dev.readthedocs.io from celery container because
that domain points to 127.0.0.1 and we don't have the storage in that
IP. So, we need to override the AZURE_MEDIA_STORAGE_HOSTNAME in the
celery container to point to the storage.

We should do this directly in `docker_compose.py` settings file, but
since we can't configure CORS in Azurite we can't do it yet.