Perform redirects at DB level
These queries a little more complex but they allow us to rely on the
database to perform the filter and check for all the redirect in a
single query than fetching all the instances of the Redirect objects
and iterating over them.

This will perform only one query to the DB and the result will be the
redirect we need to apply or no redirects at all.