Be explicit when using setup_env
Currently when `sync_repository_task` gets triggered
we use a LocalEnvironment as environment.

We want to be explicit about respecting the docker setting
in all places where we do a clone.

Also, it's really confusing depending on the order of calls to have
self.setup_env set. Wasn't able to remove it completely, we still use it
outside the function that creates the environment.