Use environment variable per run, not per container
Currently we pass the environment vars in the init method
and create the container with that.

This means that we can't modify those after the container creation.
We do allow this in the LocalBuildEnvironment.

This is needed to run git in a container (set envs like `GIT_DIR`).
And to execute the ssh agent inside, (sent envs like `SSH_AUTH_SOCKET`
and `SSH_AGENT_PID`)