Delete .cache dir on wipe
We are not deleting this directory, here is where PIP caches its
packages.

Users can still get the old (buggy) PIP version because of this.

But why wiping worked for a lot of people?

Because after wiping, we recreate the virtual environment and looks like
PIP  doesn't use the packages from the cache.
But in the next build, when reusing the virtual environment, it will use
the packages from the cache.