Skip promoting new stable if current stable is not `machine=True`
When the "stable" is managed by the user (because they have a
`origin/stable` branch, for example) we do not need to calculate which
would be the version to be promoted as stable.

So, we return immediately (without doing any computation) if the
current stable is not managed by Read the Docs.