Check for Http404 exception on infinite redirects
We can not call `proxito_404_handler` URL anymore with this setup
because we override the handler404 for these tests and we are raising
an Http404 exception with the new code. That means that the
`ServeError404` is executed twice and the second one includes an
invalid `proxito_path` argument.

We can just call the regular URL now that we raise an Http404 inside
the `ServeError404` view when there is no custom 404 page.