Refactor search view to make use of permission_classes
This is the same as https://github.com/readthedocs/readthedocs.org/pull/6133
but for the search view.

This allow us to override `get_all_projects` in .com
to filter projects using the auth backends.

And make use of the permission class from .com that makes use of the
auth backends.

This doesn't break the current search endpoint in .com
since it filters the versions using `public(user=self.request.user)`

This is required to proxy search on .com