Pull/Push cached environment using storage
Before starting clonnig the repository we check of a cached
environment in the storage. If there is one, we pull it and extract it
under `project.doc_path` (using the same structure that we have been
using) and use those cached files.

Once the build has finished and it was successful, we compress all the
environment files (pip cache, checkout, conda and virtualenv
environments) and push it to the storage.

The cached is cleaned up when the version is wiped.

This feature is under a Feature flag (CACHED_ENVIRONMENT), so we can
select specific projects to start testing out this cached
environment. It can help on projects with big/huge repositories or
that install a lot of dependencies.