Do not reset the build start time when running build env
We are creating one Setup Environment and another Build Environment.
Once each of them is created they define the `start_time` which is UTC
now. Then they use the `start_time` to calculate the `length` of the
build each time the build object is updated via the API.

Since the Build Environment calls again UTC now, we override the old
`start_time` and we loose track of the timing for the Setup
Environment (close repository, among others).

This commit saves the original `start_time` and passes it to the Build
Environment so it continues calculating the difference for the
`length` considering the Setup Environment times as well.