Report build status in a smarter way
It could happen that the person that has imported the project does not
have permissions anymore to send Build Status to the social
service (e.g. GitHub).

In that case, we were just skipping the report to the service
completely.

On the other hand, if the project didn't have a RemoreRepository, we
were trying using all the mantainers' accounts to report the status.

This commit uses the same re-trying code, but also when the first
attempt using the account of the user that imported the repository,
failed.