Add more exceptions as WARNING log level
`RepositoryError` and `ProjectConfigurationError` are user errors that
we should not log as ERROR, but as WARNING instead.

Also, the `throws` list is removed here since all these exceptions
happen inside the Build Environment who already manage them in the
`__exit__` method and so, they are never seen by the task.