Include the project slug in the PR context
We have cases where the same repo has multiple projects on RTD,
so we can support multiple builds that are happening for the same commit.

This is probably not the best design so we'll want to think more about this,
but I think showing them all for now at least makes sense.