Optimization on `sync_versions` to use ls-remote on Git VCS
When a webhook is received we synchronize the branches/tags to create new
versions under Read the Docs. For this process we do a full clone of the
repository and iterate over the branches/tags.

As this process of cloning the repository just for updating versions could be
expensive, this commit adds the ability to use `ls-remote` on Git VCS just to
query the repository for its branches/tags without performing a full clone.