Move docker limits back to setting
This setting was moved to a constant, which was harder to override and
replicated a lot of the same logic we had for settings. We hit a number
of issues around the DOCKER_LIMITS setting existing, and these constants
not being used, etc.

This hard codes a memory limit, which would be used for dev
environments outside Docker, and perhaps tests. Limits are 1g memory
and 600s build time -- up from 200m memory limit that is too small.

Production guessing of memory/time will be in ops settings now.