Feature to disable auto-generated index.md/README.rst files
When Read the Docs detects that your project doesn't have an ``index.md`` or
``README.rst``, it auto-generates one for you with instructions about how to
proceed. This behavior may be a problem for some users and they will want to
disable it.

This is happening for example when building with MkDocs and setting

```
theme:
  static_templates:
  - index.html
```

Since there is no `index.md`, Read the Docs will create one from our template
and it will override the `index.html` the user wants to show.