Domains: add ssl_status field
Currently this field depends on the provider.

For cloudflare:

- initializing
- pending_validation
- pending_issuance
- pending_deployment
- active
- pending_deletion
- deleted

For AWS:

- PENDING_VALIDATION
- SUCCESS
- ISSUED
- INACTIVE
- EXPIRED
- VALIDATION_TIMED_OUT
- REVOKED
- FAILED

I think these should be enough for all providers:

- valid
- invalid
- pending
- unknown

Also, currently we query the status via the API each time,
having them in the DB will allow us to retry the domain validation on
domains that aren't valid yet as opposed to retry on every domain as we
do now.

We do this retry when the user hits the domain list view,
whit this we can use a task instead.