Refactor VCS sync code to return RemoteRepository and RemoteOrganization
Instead of returning a list of dictionaries with different structure, we are
returning a list of RemoteRepository and RemoteOrganization. Then, we use the
same field names (`full_name` and `name`) from the objects we are going to
filter by the old values to be removed.

This reduces the complexity of choosing the right field for each service from
the dictionary previously returned; avoiding picking the wrong field, returning
`None` and removing all the Remote* objects for that user.