Add our `readthedocs_processor` data to our notifications
This fixes an issue where we were sending emails with `Contact us at {{
SUPPORT_EMAIL }}` but the support email variable was not set and the result was
an empty email.

Using our own `readthedocs_processor` here we are passing all the same variables
to the notification that to all the Django templates, avoiding this kind of confusions.