Use correct Cache-Tag (CDN) and X-RTD-Project header on subprojects
Currently, when accessing a subproject (eg. `edx-guide-for-students` being
subproject of `edx`) we are generating `Cache-Tag: edx,edx-latest`. Then, when
a build for `edx-guide-for-students` on version `latest` succeeds, we are purging
the cache tags `edx-guide-for-students-latest`.

This commit makes the `Cache-Tag` generated for subprojects to use the
subproject's slug instead of the parent's/superproject's slug and matching the
cache tag that we are purging when the build succeeds.

Besides, it fixes the `X-RTD-Project` header as well, since it was using the
same parent's/superproject's slug as value.