Reserve 1Gb for Application Memory
We have noticed some strange behavior when trying to `create_container` and
`remove_container` using the Docker API. For some reason we started getting
ReadTimeout and I'm supposing that our Docker daemon has been restarted (because
of OOM) or in a dumb state.

This commit increases from 750 to 1000 Mb of RAM memory reserved for our
application and OS processes.

The user will get 7000Mb instead of 7200Mb in build-large instances and
in build-default instances they will get 2400 intead of 2700Mb.