External versions: delete after 3 months of being merged/closed
Currently we are deleting external versions
when they are closed, but their docs remain in storage.
With this change we mark them as inactive instead.
Later a task will delete all external versions that are inactive
and their last activity was from 3 months ago.

Before being deleted the status is updated to link to the build page
instead (the build will be deleted currently, but I'm changing that in
another PR).

Ref https://github.com/readthedocs/readthedocs.org/issues/7674