Router PRs builds to last queue where a build was executed
Currently, every time a PR is triggered to be built we are building it under our
build:large queue. This is because we have no data about this version and we
build them on build:large.

This commit checks where the last build was built when a PR is triggered to be
built and use that queue to build the PR.

Closes https://github.com/readthedocs/readthedocs.org/issues/7815