Use `RemoteRepository` releation to match already imported projects
Currently production code consider a project already imported if the
`Project.repo` fuzzy matches `RemoteRepository.clone_url` even if that `Project`
is not linked to any `RemoteRepository`.

This allow anyone to import a repository manually using the real GitHub URL to
"take over" that repository and avoiding the real owner of that repository to
import it from "Import Project" page because it says:

    This repository has already been imported

and linking to the project that someone's else has imported.

With the changes on this PR, we are showing that message only if the `Project`
is connected to a `RemoteRepository` which is only possible to be done by owners
of the GitHub repository.