Adds explicit name methods to version
The version switch menu in RTD currently uses `version.slug` to display
the name of the versions. For external versions coming from PRs that
corresponds to a simple integer number.

For example, let's consider a project have 3 existing "regular
versions": v0.1, v1 and v2. In the 3rd PR submitted to GitHub,
a version with slug '3' would be created, which would result in the
version switch menu containing something like: v0.1, v1, v2, 3.
This can be slightly confusing, since it is trick to differentiate
between regular branches/tags used for versions and PR numbers,
specially when the project is starting.

This commit adds `explicit_external_name` and `explicit_name` to the
version model and uses them in the footer template to try to make a more
clear differentiation between regular builds and builds triggered by PRs.