Proxito: return user-defined HTTP headers on custom domains
We have had multiple requests from users wanting to add extra HTTP headers to
the responses on their documentation.

This commit allows this by adding a new model `projects.models.HTTPDomain` in
which we (as site administrator) can define addittional HTTP headers to a
particular domain (this objects are linked to `projects.models.Domain`).

When a request is processed, in the middleware executed by El Proxito we check
if there are extra headers for this domain and add them to the response.

Note that we currently support a pre-defined list of extra headers due a
limitation in NGINX (or in our knowledge of NGINX) that we can't include all the
upstream HTTP headers in the response and we have to set them one by one.