Track organization artifacts cleanup
Add an extra field on the `Organization` model to keep track if its artificats
were already cleanup. This allows us to filter disabled organization and run the
cleaning periodically.

The new field `Organization.artifacts_cleaned` is automatically set as `False`
when a new build happens for any Project under the organization and is set to
`True` when running a Django Admin action over the organization.