Correct an example
The example on https://docs.readthedocs.io/en/stable/guides/cross-referencing-with-sphinx.html was the opposite of what it means. It shows how to have a text distinct from the link, and as the result show the text being the link. I correct it (hopefully) by copying the example code in the result part.