Spam: allow to mark a project as (non)spam manually
This field is used from readthedocs-ext to always return 0 as score when the
project was manually marked as `is_spam=False`. This way, even if there are some
`SpamRule`s that apply to the project, we can manually tell our system that this
project is valid and remove false positives.

See https://github.com/readthedocs/readthedocs-ext/issues/458