Config file: system_site_packages overwritten from project's setting
I created a test case to demostrate what was reported in the issue. Summarizing,
if the project uses a config file and does not define `python.system_packages`
but has `Project.use_system_packages=True` the config option is `True` which is
wrong. When a config file is used, settings at Project level should be ignored
as this messages from "Advanced settings" states:

> These settings can be configured using a configuration file. That's the
recommended way to set up your project. Settings listed here are ignored when
using a configuration file.

Reference #8782