Test: LocalBuildEnvironment run(), __enter__ and __exit__ required
We use this environment a lot from our tests, which I don't like. However, for
now, I'm keeping the because the change is not small.

We should split unit-test from integration test and reduce the complexity of
executing "real commands" when calling tests in exchange of more mocked data
instead and unit tests for the logic.