Cleanup: remove "Wipe" feature
We are always using `CLEAN_AFTER_BUILD=True`, so there is no environment to wipe
anymore. This is a confusing message to send to the user since it has no effect.

I'm also removing the code because it doesn't make sense anymore. If we ever
need it again, we can re-implement it correctly and recover it from here.