Test: comments and notes about what how to migrate them
There are lot of tests not really accurate and very complex without reason. I'd
like to use the same flow in production than in tests instead of
short-circuiting/hacking the process in tests to behave as we want.

We should use mocks to avoid executing commands when running tests instead, but
making sure the "build process" is following exact the same flow in both
environments. If we want to test something in particular outside the flow, we
should create unit tests for those functions/methods.