django-storages: upgrade
Currently, the development environment is broken. We haven't noticed this
because it fails on creating the `staticfiles.json` file on `--init` and without
that the whole environment is broken.

By deleting my local volumes I was able to reproduce it. Now, my environment is
broken and it does not work without this patch.

Note that this commit requires a change in `common/` repository to auto-create
the buckets on S3 (MinIO) when called with `--init`. This is required because
newer version of django-storages removed support for auto-creation.

Closes #8812