Docs: reduce visibility of Config File V1
People is still arriving at Config File V1 and we want to avoid that as much as
possible. However, we can't yet remove this documentation page since it's still
working and there are people already using it.

We should find a way to migrate these projects off from Config File V1 and
finally remove this documentation page. However, that requires a much bigger
amount of work.