Reduce verbose logging on generic command failure
We updated the copy used for generic errors with the celery refactor,
but also converted a generic command failure to `BuildUserError(msg)`.
This caused the command output to always be used as the error message
shown to the user.

More error classes is probably the direction we want to go here, perhaps
even going as far as to provide more helpful errors for each step that
commonly emits the generic user error class.