Sentry: ignore logging known exceptions
By reading
https://docs.sentry.io/clients/python/integrations/#additional-settings it seems
that we should be able to ignore known exceptions to be sent to Sentry. Note
that we have this setting already defined at
https://github.com/readthedocs/readthedocs.org/blob/b5908b0fc65cb02fc840b23eeb4218b3115b6590/readthedocs/settings/base.py#L453

However, we need to add these known exceptions to
https://github.com/readthedocs/readthedocs.org/blob/b5908b0fc65cb02fc840b23eeb4218b3115b6590/readthedocs/projects/tasks/builds.py#L224-L229

This commit adds the known exceptions to be ignored by Celery and not sent to
Sentry either.