Build: use `ubuntu-20.04` image for setup VCS step
We were _always_ using `readthedocs/build:latest` to run all the VCS steps (e.g.
`git clone`). That image had an old `ca-certificates` Ubuntu package that does
not contains the newer Lets Encrypt certificate and because of that many git
command failed.

Now, we had rebuilt the `ubuntu-20.04` image which includes a newer version of
this package and we are defining this image to be used by default in the setup
VCS step. The main goal is to avoid calling `apt-get install ca-certificate` on
each build and reduce this overhead.