Build: allow base environment to receive `**kwargs`
We need this because there are some attributes that are passed only to
DockerBuildEnvironment but are not allowed in LocalBuildEnvironment (e.g.
`container_image`).

As LocalBuildEnvironment is used in the tests, they were failing because it
doesn't accept this argument.

Again, we should refactor our test's code to completely remove
LocalBuildEnvironment and rely 100% on DockerBuildEnvironment.