Test: workaround mocked instance
`task.id` is a mock.Mock object and it's not possible to save it into the
database. I've tried using `.return_value` and others but I failed when trying
to mock it and return a valid str/int.

So, for now, we can keep this hack here and only save the value into the
database if it's a str/int.