Build: use same build environment for setup and build
We can't do `with self.build_environment:` twice because the context manager
kills the Docker container on `__exit__`. So, the second time we call the `with`
it will create a new container which won't have all the work done in the
previous steps.

We could avoid this by using shared volumes instead, but that's a bigger
refactor. For now, I'm moving the part that creates the build environment back
to the Celery task and executing `setup_environment` _and_ `build` steps from
there, which is not ideally, but solves the problem for now.