SyncRepositoryTask: rate limit to 1 per minute per project
Different reasons (e.g. multiple push) makes our platform to trigger
`SyncRepositoryTask` multiple times one next to the other for the same project.
This consumes a lot of resources (clone the repository, query the databse,
update the database's versions, etc) but in the end it's is useless because
there is nothing has changed so quickly in the repository to require re-syncing
the versions.

We've found projects triggernig more than 1.5k of this task in a single day or
even in less time. This has made our database's CPU to spike and reduce the
availability of our workers to execute `UpdateDocsTask`, blocking our users and
producing a worse UX to them.

This commit makes usage of a lock (based on Celery's `memcache_lock`) to execute
up to 1 task per minute per project, and discard the rest of them.

Closes https://github.com/readthedocs/readthedocs.org/issues/8922