Build: do not send VCS build status on specific exceptions
When the exception is an exception we know it does not require communicating the
VCS service about its state, we avoid sending the build status to the VCS
service.

This is currently used to avoid sending a "failure" (red cross) notification
when the build is marked as duplicated. In the future, it could also be used
when the build is auto-cancelled because a new commit on the same branch was
done (see https://github.com/readthedocs/readthedocs.org/issues/8961)

Closes #9040