Build: do not upload `build.tool` to production S3 (#9098)
* Build: do not upload `build.tool` to production S3

Avoid uploading the .tar.gz to production S3 since we will be managing this via
CircleCI and an AWS orbs. See
https://github.com/readthedocs/readthedocs-docker-images/pull/179

We need to keep the upload part for development environment, tho.

* Build: join TOOL and VERSION with a `-` to simplify CircleCI routine

* Do not delete .tar.gz from local host when on production

* Remove old documentation

We are not going to run this command in production by hand anymore. However, in
case it's required, I'm linking to the issue that explains how to do it.

* Skip uploading when running from inside CIRCLECI

* Split docker iamge name from full to just keep the name without date

* Revert "Split docker iamge name from full to just keep the name without date"

This reverts commit e08a7de89869f7b172a6fe792e235484d24e2cdc.

* Receive arguments as previously

* Comment about pinning transitive dependencies

* Build Python 3.6.15 using `clang`