Mamba: remove CONDA_USES_MAMBA feature flag
Mamba can be used via `build.tools.python: mambaforge` now.

This PR removes the `CONDA_USES_MAMBA` feature flag since we already contact
most of the users and opened a PR on their repositories with the config file
required for this change and we also sent an in-site notification to all those
projects still using it.

The message said this feature flag will be removed on May 1st. So, we are all
good to merge and deploy next week (May 3rd).