Tests: make `tests-embedapi` require regular `tests` first
I'm adding this requirement so we don't trigger all the tests concurrently and
also avoid running EmbedAPI tests if there is a failure on any of the other two.
This way we will be saving some resources here.

EmbedAPI tests are _not required_ to merge a PR, so I think it's fine if they
are not always run.